backend/requirements.txt
# python 3.12.0
# Flask Core
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flask_login==0.6.3

# DB
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2==2.9.10

# env
python-dotenv==1.0.1
